ui/handler, encoder: Allow overriding lag in frames
The lag in frames is not dictated by the number of threads being used by some encoders. At least for hardware encoders, the expected frame lag in real time encoding is 1 + the number of bframes

+++ source/encoder.hpp
@@ -94,8 +94,8 @@ namespace obsffmpeg {
 		ffmpeg::swscale _swscale;
 		AVPacket        _current_packet;
 
-		int64_t _lag_in_frames;
-		int64_t _count_send_frames;
+		size_t _lag_in_frames;
+		size_t _count_send_frames;
 
 		// Extra Data
 		bool                 _have_first_frame;
